split file into handler and add toml step
