+++ packages/cli/src/commands/experimental/studio.js
@@ -1,11 +1,9 @@
 import terminalLink from 'terminal-link'
 
-import { isModuleInstalled, installRedwoodModule } from '../../lib/packages'
-
 export const command = 'studio'
 export const description = 'Run the Redwood development studio'
 
-export const builder = (yargs) => {
+export function builder(yargs) {
   yargs.epilogue(
     `Also see the ${terminalLink(
       'Redwood CLI Reference',
@@ -14,23 +12,7 @@ export const builder = (yargs) => {
   )
 }
 
-export const handler = async () => {
-  try {
-    // Check the module is installed
-    if (!isModuleInstalled('@redwoodjs/studio')) {
-      console.log(
-        'The studio package is not installed, installing it for you, this may take a moment...'
-      )
-      await installRedwoodModule('@redwoodjs/studio')
-      console.log('Studio package installed successfully.')
-    }
-
-    // Import studio and start it
-    const { start } = await import('@redwoodjs/studio')
-    await start()
-  } catch (e) {
-    console.log('Cannot start the development studio')
-    console.log(e)
-    process.exit(1)
-  }
+export async function handler(options) {
+  const { handler } = await import('./studioHandler')
+  return handler(options)
 }
